editando medida

+++ database/migrations/2020_11_20_120348_create_medida_table.php
@@ -22,9 +22,9 @@ class CreateMedidaTable extends Migration
             $table->string('tipo')->nullable();
             $table->text('texto_noticia')->nullable();
             $table->decimal('diferenca_financas')->nullable();
-            $table->decimal('diferenca_popularidade_empresarios')->nullable();
-            $table->decimal('diferenca_popularidade_trabalhadores')->nullable();
-            $table->decimal('diferenca_popularidade_estado')->nullable();
+            $table->integer('diferenca_popularidade_empresarios')->nullable();
+            $table->integer('diferenca_popularidade_trabalhadores')->nullable();
+            $table->integer('diferenca_popularidade_estado')->nullable();
             $table->timestamps();
         });
     }
